+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- In database create, parse numeric string values into strings, E.g., Postgres version property will be now correctly parsed as string value.
+
 ## [3.20.0] - 2025-05-26
 
 ### Added